Hint that device may be locked if connection fails
I spent a few hours trying to figure out why I could not connect to my new Broadlink product before realising that I had not unlocked it. Providing a hint to users that the reason they may not be able to connect could be because the device is locked could save a few headaches.